# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VisionTrack - Local Video Analytics Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Video processing jobs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      description: Returns all video processing jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  /upload:
    post:
      operationId: uploadVideo
      tags: [jobs]
      summary: Upload a video file
      description: Upload a video file and create a new job
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /process/{jobId}:
    post:
      operationId: processJob
      tags: [jobs]
      summary: Trigger video processing
      description: Start the detection pipeline for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /status/{jobId}:
    get:
      operationId: getJobStatus
      tags: [jobs]
      summary: Get job status
      description: Get the current status of a processing job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /download/{jobId}:
    get:
      operationId: downloadReport
      tags: [jobs]
      summary: Download Excel report
      description: Download the generated Excel analytics report for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "404":
          description: Report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /jobs/{jobId}:
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      description: Delete a job and its associated files
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Job deleted
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Job:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        status:
          type: string
          enum: [pending, processing, complete, failed]
        fileSizeBytes:
          type: ["integer", "null"]
        durationSeconds:
          type: ["number", "null"]
        totalDetections:
          type: ["integer", "null"]
        totalTracks:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        reportPath:
          type: ["string", "null"]
        annotatedVideoPath:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - filename
        - originalName
        - status
        - fileSizeBytes
        - durationSeconds
        - totalDetections
        - totalTracks
        - errorMessage
        - reportPath
        - annotatedVideoPath
        - createdAt
        - updatedAt

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error