# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VisionTrack - Local Video Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from "./jobStatus";

export interface Job {
  id: number;
  filename: string;
  originalName: string;
  status: JobStatus;
  /** @nullable */
  fileSizeBytes: number | null;
  /** @nullable */
  durationSeconds: number | null;
  /** @nullable */
  totalDetections: number | null;
  /** @nullable */
  totalTracks: number | null;
  /** @nullable */
  errorMessage: string | null;
  /** @nullable */
  reportPath: string | null;
  /** @nullable */
  annotatedVideoPath: string | null;
  createdAt: Date;
  updatedAt: Date;
}